+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 * **m**aintenance
 * **p**erformance
 
+### 2020.8.7
+
+* **f** Save first tagged argument in arg property
+* **f** Save all key value arguments in args property
+* **f** Save all untagged arguments in plain property
+
 ### 2020.8.6
 
 * **f** `--` marks all subsequent inputs as arguments